+++ azure-test/tests/azure_log_alert/test-hydrate-query.sql
@@ -1,0 +1,3 @@
+select name, akas, title
+from azure.azure_log_alert
+where name = '{{ resourceName }}' and resource_group = '{{ resourceName }}';