+++ azure-test/tests/azure_log_alert/variables.tf
@@ -1,0 +1,95 @@
+variable "resource_name" {
+  type        = string
+  default     = "turbot-test-20200125-create-update"
+  description = "Name of the resource used throughout the test."
+}
+
+variable "azure_environment" {
+  type        = string
+  default     = "public"
+  description = "Azure environment used for the test."
+}
+
+variable "azure_subscription" {
+  type        = string
+  default     = "3510ae4d-530b-497d-8f30-53b9616fc6c1"
+  description = "Azure subscription used for the test."
+}
+
+provider "azurerm" {
+  features {}
+  environment     = var.azure_environment
+  subscription_id = var.azure_subscription
+}
+
+data "azurerm_client_config" "current" {}
+data "null_data_source" "resource" {
+  inputs = {
+    scope = "azure:///subscriptions/${data.azurerm_client_config.current.subscription_id}"
+  }
+}
+
+resource "azurerm_resource_group" "named_test_resource" {
+  name     = var.resource_name
+  location = "West US"
+}
+
+resource "azurerm_monitor_action_group" "named_test_resource" {
+  name                = var.resource_name
+  resource_group_name = azurerm_resource_group.named_test_resource.name
+  short_name          = "p0action"
+
+  webhook_receiver {
+    name        = "callmyapi"
+    service_uri = "http://example.com/alert"
+  }
+}
+
+resource "azurerm_storage_account" "named_test_resource" {
+  name                     = var.resource_name
+  resource_group_name      = azurerm_resource_group.named_test_resource.name
+  location                 = azurerm_resource_group.named_test_resource.location
+  account_tier             = "Standard"
+  account_replication_type = "LRS"
+}
+
+resource "azurerm_monitor_activity_log_alert" "named_test_resource" {
+  name                = var.resource_name
+  resource_group_name = azurerm_resource_group.named_test_resource.name
+  scopes              = [azurerm_resource_group.named_test_resource.id]
+  description         = "This alert will monitor a specific storage account updates."
+
+  criteria {
+    resource_id    = azurerm_storage_account.named_test_resource.id
+    operation_name = "Microsoft.Storage/storageAccounts/write"
+    category       = "Recommendation"
+  }
+
+  action {
+    action_group_id = azurerm_monitor_action_group.named_test_resource.id
+
+    webhook_properties = {
+      from = "terraform"
+    }
+  }
+}
+
+output "resource_aka" {
+  value = "azure://${azurerm_monitor_activity_log_alert.named_test_resource.id}"
+}
+
+output "resource_aka_lower" {
+  value = "azure://${lower(azurerm_monitor_activity_log_alert.named_test_resource.id)}"
+}
+
+output "resource_name" {
+  value = var.resource_name
+}
+
+output "resource_id" {
+  value = azurerm_monitor_activity_log_alert.named_test_resource.id
+}
+
+output "subscription_id" {
+  value = var.azure_subscription
+}